+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/BugsnagPerformance.kt
@@ -6,7 +6,6 @@ import android.content.Context
 import android.net.Uri
 import android.os.SystemClock
 import com.bugsnag.android.performance.BugsnagPerformance.start
-import com.bugsnag.android.performance.internal.controls.CompositeSpanControlProvider
 import com.bugsnag.android.performance.controls.SpanQuery
 import com.bugsnag.android.performance.internal.Connectivity
 import com.bugsnag.android.performance.internal.DiscardingSampler
@@ -22,10 +21,12 @@ import com.bugsnag.android.performance.internal.SamplerTask
 import com.bugsnag.android.performance.internal.SendBatchTask
 import com.bugsnag.android.performance.internal.Task
 import com.bugsnag.android.performance.internal.Worker
+import com.bugsnag.android.performance.internal.controls.CompositeSpanControlProvider
 import com.bugsnag.android.performance.internal.createResourceAttributes
 import com.bugsnag.android.performance.internal.integration.NotifierIntegration
 import com.bugsnag.android.performance.internal.isInForeground
 import com.bugsnag.android.performance.internal.metrics.SystemConfig
+import com.bugsnag.android.performance.internal.plugins.PluginManager
 import com.bugsnag.android.performance.internal.processing.ImmutableConfig
 import java.net.URL
 
@@ -78,7 +79,7 @@ public object BugsnagPerformance {
         if (!isStarted) {
             synchronized(this) {
                 if (!isStarted) {
-                    startUnderLock(ImmutableConfig(configuration))
+                    startUnderLock(configuration)
                     isStarted = true
                 }
             }
@@ -91,8 +92,13 @@ public object BugsnagPerformance {
         Logger.w("BugsnagPerformance.start has already been called")
     }
 
-    private fun startUnderLock(configuration: ImmutableConfig) {
-        Logger.delegate = configuration.logger
+    private fun startUnderLock(externalConfiguration: PerformanceConfiguration) {
+        Logger.delegate = ImmutableConfig.getLogger(externalConfiguration)
+
+        val pluginManager = PluginManager(externalConfiguration.plugins)
+        pluginManager.installPlugins(externalConfiguration)
+
+        val configuration = ImmutableConfig(externalConfiguration, pluginManager)
         val tracer = instrumentedAppState.configure(configuration)
 
         if (configuration.autoInstrumentAppStarts) {
@@ -173,6 +179,9 @@ public object BugsnagPerformance {
             workerTasks.add(SendBatchTask(delivery, tracer, resourceAttributes))
             workerTasks.add(RetryDeliveryTask(persistence.retryQueue, httpDelivery, connectivity))
 
+            // starting plugins is the last thing to do before starting the first tasks
+            pluginManager.startPlugins()
+
             return@Worker workerTasks
         }
 
@@ -180,6 +189,10 @@ public object BugsnagPerformance {
         tracer.worker = bsgWorker
 
         loadModules()
+
+        spanControlProvider.addProviders(
+            pluginManager.completeContext?.spanControlProviders.orEmpty(),
+        )
 
         bsgWorker.start()